chore: migrate shared package to Swan
