+++ swan/src/shared/get-label-from-slot.ts
@@ -10,14 +10,25 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-/**
- * ⚠️ IMPORTANT: SOURCE OF TRUTH MOVED ⚠️
- *
- * The authoritative implementation is now in:
- * swan/src/shared/get-label-from-slot.ts
- *
- * This file is maintained for backward compatibility ONLY.
- * DO NOT modify this file directly; instead make changes in Swan.
- */
-
-export * from '@spectrum-web-components/swan/shared/get-label-from-slot.js';
+export const getLabelFromSlot = (
+    label: string,
+    slotEl: HTMLSlotElement
+): string | null => {
+    if (label) {
+        return null;
+    }
+    const textContent = slotEl
+        .assignedNodes()
+        .reduce((accumulator: string, node: Node) => {
+            if (node.textContent) {
+                return accumulator + node.textContent;
+            } else {
+                return accumulator;
+            }
+        }, '');
+    if (textContent) {
+        return textContent.trim();
+    } else {
+        return null;
+    }
+};
