+++ tools/shared/src/platform.ts
@@ -10,62 +10,14 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-function testUserAgent(re: RegExp): boolean {
-    return typeof window !== 'undefined' && window.navigator != null
-        ? re.test(window.navigator.userAgent)
-        : false;
-}
+/**
+ * ⚠️ IMPORTANT: SOURCE OF TRUTH MOVED ⚠️
+ *
+ * The authoritative implementation is now in:
+ * swan/src/shared/platform.ts
+ *
+ * This file is maintained for backward compatibility ONLY.
+ * DO NOT modify this file directly; instead make changes in Swan.
+ */
 
-function testPlatform(re: RegExp): boolean {
-    return typeof window !== 'undefined' && window.navigator != null
-        ? re.test(window.navigator.platform)
-        : false;
-}
-
-/* c8 ignore next 3 */
-export function isMac(): boolean {
-    return testPlatform(/^Mac/);
-}
-
-export function isIPhone(): boolean {
-    return testPlatform(/^iPhone/);
-}
-
-export function isIPad(): boolean {
-    return (
-        testPlatform(/^iPad/) ||
-        // iPadOS 13 lies and says it's a Mac, but we can distinguish by detecting touch support.
-        (isMac() && navigator.maxTouchPoints > 1)
-    );
-}
-
-export function isIOS(): boolean {
-    return isIPhone() || isIPad();
-}
-
-/* c8 ignore next 3 */
-export function isAppleDevice(): boolean {
-    return isMac() || isIOS();
-}
-
-/* c8 ignore next 3 */
-export function isWebKit(): boolean {
-    return testUserAgent(/AppleWebKit/) && !isChrome();
-}
-
-/* c8 ignore next 3 */
-export function isChrome(): boolean {
-    return testUserAgent(/Chrome/);
-}
-
-export function isFirefox(): boolean {
-    return testUserAgent(/Firefox/) && !isSeamonkey();
-}
-
-export function isSeamonkey(): boolean {
-    return testUserAgent(/Seamonkey/);
-}
-
-export function isAndroid(): boolean {
-    return testUserAgent(/Android/);
-}
+export * from '@spectrum-web-components/swan/shared/platform.js';
