+++ tools/shared/src/reparent-children.ts
@@ -9,81 +9,15 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-function restoreChildren<T extends Element>(
-    placeholderItems: Comment[],
-    srcElements: T[],
-    cleanupCallbacks: ((el: T) => void)[] = []
-): T[] {
-    for (let index = 0; index < srcElements.length; ++index) {
-        const srcElement = srcElements[index];
-        const placeholderItem = placeholderItems[index];
-        const parentElement =
-            placeholderItem.parentElement || placeholderItem.getRootNode();
-        if (cleanupCallbacks[index]) {
-            cleanupCallbacks[index](srcElement);
-        }
-        if (parentElement && parentElement !== placeholderItem) {
-            parentElement.replaceChild(srcElement, placeholderItem);
-        }
-        delete placeholderItems[index];
-    }
-    return srcElements;
-}
 
-export const reparentChildren = <T extends Element>(
-    srcElements: T[],
-    destination: Element,
-    {
-        position,
-        prepareCallback,
-    }: {
-        position: InsertPosition;
-        prepareCallback?: (el: T) => ((el: T) => void) | void;
-    } = { position: 'beforeend' }
-): (() => T[]) => {
-    let { length } = srcElements;
-    if (length === 0) {
-        return () => srcElements;
-    }
+/**
+ * ⚠️ IMPORTANT: SOURCE OF TRUTH MOVED ⚠️
+ *
+ * The authoritative implementation is now in:
+ * swan/src/shared/reparent-children.ts
+ *
+ * This file is maintained for backward compatibility ONLY.
+ * DO NOT modify this file directly; instead make changes in Swan.
+ */
 
-    let step = 1;
-    let index = 0;
-
-    if (position === 'afterbegin' || position === 'afterend') {
-        step = -1;
-        index = length - 1;
-    }
-
-    const placeholderItems = new Array<Comment>(length);
-    const cleanupCallbacks = new Array<(el: T) => void>(length);
-    const placeholderTemplate: Comment = document.createComment(
-        'placeholder for reparented element'
-    );
-
-    do {
-        const srcElement = srcElements[index];
-        if (prepareCallback) {
-            cleanupCallbacks[index] = prepareCallback(srcElement) as (
-                el: T
-            ) => void;
-        }
-        placeholderItems[index] = placeholderTemplate.cloneNode() as Comment;
-
-        const parentElement =
-            srcElement.parentElement || srcElement.getRootNode();
-        if (parentElement && parentElement !== srcElement) {
-            parentElement.replaceChild(placeholderItems[index], srcElement);
-        }
-        destination.insertAdjacentElement(position, srcElement);
-
-        index += step;
-    } while (--length > 0);
-
-    return function (): T[] {
-        return restoreChildren<T>(
-            placeholderItems,
-            srcElements,
-            cleanupCallbacks
-        );
-    };
-};
+export * from '@spectrum-web-components/swan/shared/reparent-children.js';
