+++ tsconfig.json
@@ -24,6 +24,7 @@
         "resolveJsonModule": true,
         "paths": {
             "@spectrum-web-components/swan/base/*": ["../swan/src/base/*"],
+            "@spectrum-web-components/swan/shared/*": ["../swan/src/shared/*"],
             "@spectrum-web-components/swan/*": ["../swan/src/components/*"]
         }
     },
